+++ packages/user-mgmt/.dev.vars-example
@@ -1,0 +1,1 @@
+EMAIL_DKIM_PRIVATE_KEY = "YOUR_BASE_64_ENCODED_PRIVATE_KEY_GOES_HERE"
