+++ packages/user-mgmt/wrangler.toml
@@ -15,6 +15,9 @@ database_id = "08d1a683-e910-43c4-bce1-918319e23e7f"
 [vars]
 EMAIL_FROM = "noreply@justblackmagic.com"
 EMAIL_FROM_NAME = "Black Magic Consulting"
+
+EMAIL_DKIM_SELECTOR = "mailchannels._domainkey"
+EMAIL_DKIM_DOMAIN = "justblackmagic.com"
 FORGOT_PASSWORD_URL = "https://account-pages.tools.justblackmagic.com/forgot-password-reset.html"
 TOKEN_VALID_MINUTES = 60
